backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.27.0
websockets==12.0
openai==1.23.6
supabase==2.4.0
